+++ MHD/C_src/KEMO_GL/modify_object_4_viewer_c.h
@@ -21,7 +21,7 @@
 void cal_range_4_mesh_c(struct viewer_mesh *mesh_s, struct view_element *view);
 void modify_object_multi_viewer_c(double dist, struct viewer_mesh *mesh_s);
 void set_axis_positions(struct view_element *view, GLfloat dist, 
-                        GLfloat *axis_delta, GLfloat *axis_org);
+                        double *axis_delta, double *axis_org);
 
 void cal_psf_viewer_range(struct psf_data **psf_s, struct kemo_array_control *psf_a,  
                           struct psf_data *fline_s, struct fline_menu_val *fline_m,
