+++ MHD/C_src/KEMO_GL/set_PSF_patches_to_buf.c
@@ -115,8 +115,8 @@ int count_psf_arrows_to_buf(int ncorner, struct psf_data *psf_s, struct psf_menu
 
 int set_psf_arrows_to_buf(int ist_patch, int ncorner, struct psf_data *psf_s, struct psf_menu_val *psf_m, 
 			struct gl_strided_buffer *strided_buf) {
-	float x_line[6], dir_line[6], color_line[8];
-	float xyz[18*ncorner], nor[18*ncorner], col[24*ncorner];
+	double x_line[6], dir_line[6], color_line[8];
+	double xyz[18*ncorner], nor[18*ncorner], col[24*ncorner];
 	GLdouble dcolor[4];
 	int num_wall, inum_buf;
 	
@@ -125,7 +125,7 @@ int set_psf_arrows_to_buf(int ist_patch, int ncorner, struct psf_data *psf_s, st
 	int inod, i, k, nd;
 	
 	int icomp = psf_s->istack_comp[psf_m->if_draw_psf];
-	float radius = (float) psf_m->vector_thick;
+	double radius = psf_m->vector_thick;
 	double ascale = ONE / psf_m->scale_vect;
 	
 	inum_buf = ist_patch;
